Move backend files to `src` directory
I'm about to add some more files so it's best to keep them organised.
