+++ src/app.js
@@ -6,7 +6,7 @@ var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 
-var config = require('./config');
+var config = require('../config');
 
 var domain = 'localhost';
 
@@ -22,12 +22,12 @@ switch (config.appsettings.env) {
 }
 
 process.env.HOST = domain;
-var routes = require('./routes/api');
+var routes = require('../routes/api');
 
 var app = express();
 
 // view engine setup
-app.set('views', path.join(__dirname + '/public', 'views'));
+app.set('views', path.join(__dirname, '..', 'public', 'views'));
 app.set('view engine', 'jade');
 
 // uncomment after placing your favicon in /public
@@ -36,8 +36,8 @@ app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(require('stylus').middleware(path.join(__dirname, 'public')));
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(require('stylus').middleware(path.join(__dirname, '..', 'public')));
+app.use(express.static(path.join(__dirname, '..', 'public')));
 
 //app.use('/', routes);
 app.use('/api', routes);
